+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Avoid showing empty chats in resume. #326
 - Fix anthropic adaptive thinking option causing invalid request errors.
+- Fix anthropic web server search corner case causing invalid request errors.
 
 ## 0.104.1